requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
requests>=2.28.0
tqdm>=4.65.0
numpy>=1.24.0
